+++ SHM/src/main/java/ua/com/manometer/dao/CustomerDAO.java
@@ -9,9 +9,9 @@ public interface CustomerDAO {
 
 	public List<Customer> listCustomer();
 
-	public void removeCustomer(Long id);
+	public void removeCustomer(Integer id);
 
-    public Customer getCustomer(Long id);
+    public Customer getCustomer(Integer id);
 
     public List<String> findByShortNameExample(String customerTemplate);
