+++ SHM/src/main/java/ua/com/manometer/service/CurrencyService.java
@@ -10,6 +10,6 @@ public interface CurrencyService {
 
 	public List<Currency> listCurrency();
 
-	public void removeCurrency(Long id);
+	public void removeCurrency(Integer id);
 
 }
